third_party/WebKit/LayoutTests/fast/regions/webkit-named-flow-collection-empty-getters.html

<!doctype html>
<html>
    <body>
        <script src="../../resources/js-test.js"></script>
        <script>
            description("Test for <a href='https://code.google.com/p/chromium/issues/detail?id=259339'>[259339] NamedFlowCollection getters should follow the same pattern as HTMLCollection</a>");

            var namedFlowCollection = document.webkitGetNamedFlows();
            shouldBeUndefined("namedFlowCollection[0]");
            shouldBeUndefined("namedFlowCollection['flow']");
            shouldBeNull("namedFlowCollection.item(0)");
            shouldBeNull("namedFlowCollection.namedItem('flow')");
        </script>
    </body>
</html>
